+++ app/__init__.py
@@ -1,3 +1,3 @@
 __all__ = ["main"]
 
-__version__ = "1.0.0"
+__version__ = "1.0.1"